fix(nx-rt-core): correct runtime startup and service session lifetimes (#106)
Startup recorded the kernel-supplied main-thread handle only while parsing a loader configuration block, and a service manager replaced its session on a second open, so one caller could close it out from under another.

- Take the main-thread handle as an `init_once` argument, so a launch with a null configuration block records it too rather than aborting in the first startup step that reads it back
- Own the applet message event for the life of the session: the Application Manager mints a fresh copy handle per request, so the accessors leaked one per call and handed the C surface a handle it would close as well. Adds the matching `close_handle` to `nx_svc::sync`, which was the only resource module without one
- Count session owners in `applet`, `apm`, `hid`, `set` and `time`, as `nv` and `vi` already did, so a second `init` joins the open session and only the last `exit` closes it
- Give the syscall hints and the applet-type conversion total forms, retiring an abort reachable from the C hint query, an unreachable error variant, and the crate's only `transmute`
- Refuse display names the C structure cannot carry, narrow the `nvmap` kind fallibly, read the firmware version through the manager that chose the session protocol, and select `rt-link` on the staticlib that is actually linked

+++ subprojects/nx-rt-nro/src/ffi/libnx/applet.rs
@@ -45,5 +45,5 @@ pub unsafe extern "C" fn __nx_rt_nro__libnx_applet_initialize() -> u32 {
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn __nx_rt_nro__libnx_applet_get_applet_type() -> i32 {
     // Report the loader-supplied applet type from the parsed environment state.
-    env::applet_type().as_raw() as i32
+    env::applet_type().as_raw()
 }
